Fixed imports after moving Player.
